feat: display ckBTC transactions
