+++ frontend/src/lib/components/accounts/CkBTCTransactionsList.svelte
@@ -3,9 +3,13 @@
   import { onMount } from "svelte";
   import { loadCkBTCAccountNextTransactions } from "$lib/services/ckbtc-transactions.services";
   import { IcrcTransactionData } from "$lib/types/transaction";
-  import { sortTransactions } from "$lib/utils/icrc-transactions.utils";
   import { icrcTransactionsStore } from "$lib/stores/icrc-transactions.store";
   import { CKBTC_INDEX_CANISTER_ID } from "$lib/constants/canister-ids.constants";
+  import {
+    getSortedTransactionsFromStore,
+    isIcrcTransactionsCompleted,
+  } from "$lib/utils/icrc-transactions.utils";
+  import IcrcTransactionsList from "$lib/components/accounts/IcrcTransactionsList.svelte";
 
   export let account: Account;
 
@@ -14,7 +18,7 @@
   const loadTransactions = async () => {
     loading = true;
     await loadCkBTCAccountNextTransactions({
-      account
+      account,
     });
     loading = false;
   };
@@ -22,8 +26,18 @@
   onMount(async () => await loadTransactions());
 
   let transactions: IcrcTransactionData[];
-  $: transactions = sortTransactions(
-    $icrcTransactionsStore[CKBTC_INDEX_CANISTER_ID.toText()]?.[account.identifier]
-      ?.transactions
-  );
+  $: transactions = getSortedTransactionsFromStore({
+    store: $icrcTransactionsStore,
+    canisterId: CKBTC_INDEX_CANISTER_ID,
+    account,
+  });
+
+  let completed: boolean;
+  $: completed = isIcrcTransactionsCompleted({
+    store: $icrcTransactionsStore,
+    canisterId: CKBTC_INDEX_CANISTER_ID,
+    account,
+  });
 </script>
+
+<IcrcTransactionsList {account} {transactions} {loading} {completed} />
